Split address.h/address.c into platform files
