+++ platform/posix/address.c
@@ -1,0 +1,65 @@
+#include "address.h"
+
+int
+coap_address_isany(const coap_address_t *a) {
+  /* need to compare only relevant parts of sockaddr_in6 */
+  switch (a->addr.sa.sa_family) {
+  case AF_INET:
+    return a->addr.sin.sin_addr.s_addr == INADDR_ANY;
+  case AF_INET6:
+    return memcmp(&in6addr_any,
+                  &a->addr.sin6.sin6_addr,
+                  sizeof(in6addr_any)) == 0;
+  default:
+    ;
+  }
+
+  return 0;
+}
+
+int
+coap_is_mcast(const coap_address_t *a) {
+  if (!a)
+    return 0;
+
+ switch (a->addr.sa.sa_family) {
+ case AF_INET:
+   return IN_MULTICAST(a->addr.sin.sin_addr.s_addr);
+ case  AF_INET6:
+   return IN6_IS_ADDR_MULTICAST(&a->addr.sin6.sin6_addr);
+ default:  /* fall through and signal error */
+   ;
+  }
+ return 0;
+}
+
+int
+coap_address_equals(const coap_address_t *a, const coap_address_t *b) {
+  assert(a); assert(b);
+
+  if (a->size != b->size || a->addr.sa.sa_family != b->addr.sa.sa_family)
+    return 0;
+
+  /* need to compare only relevant parts of sockaddr_in6 */
+ switch (a->addr.sa.sa_family) {
+ case AF_INET:
+   return a->addr.sin.sin_port == b->addr.sin.sin_port &&
+     memcmp(&a->addr.sin.sin_addr, &b->addr.sin.sin_addr,
+     sizeof(struct in_addr)) == 0;
+ case AF_INET6:
+   return a->addr.sin6.sin6_port == b->addr.sin6.sin6_port &&
+     memcmp(&a->addr.sin6.sin6_addr, &b->addr.sin6.sin6_addr,
+     sizeof(struct in6_addr)) == 0;
+ default: /* fall through and signal error */
+   ;
+ }
+ return 0;
+}
+
+void
+coap_address_init(coap_address_t *addr) {
+  assert(addr);
+  memset(addr, 0, sizeof(coap_address_t));
+  addr->size = sizeof(addr->addr);
+}
+
