--- src/address.c
@@ -1,48 +1,0 @@
-/* address.c -- representation of network addresses
- *
- * Copyright (C) 2015 Olaf Bergmann <bergmann@tzi.org>
- *
- * This file is part of the CoAP library libcoap. Please see
- * README for terms of use.
- */
-
-#ifdef WITH_POSIX
-#include <assert.h>
-#include <netinet/in.h>
-#include <sys/socket.h>
-
-#include "address.h"
-
-int 
-coap_address_equals(const coap_address_t *a, const coap_address_t *b) {
-  assert(a); assert(b);
-
-  if (a->size != b->size || a->addr.sa.sa_family != b->addr.sa.sa_family)
-    return 0;
-  
-  /* need to compare only relevant parts of sockaddr_in6 */
- switch (a->addr.sa.sa_family) {
- case AF_INET:
-   return 
-     a->addr.sin.sin_port == b->addr.sin.sin_port && 
-     memcmp(&a->addr.sin.sin_addr, &b->addr.sin.sin_addr, 
-	    sizeof(struct in_addr)) == 0;
- case AF_INET6:
-   return a->addr.sin6.sin6_port == b->addr.sin6.sin6_port && 
-     memcmp(&a->addr.sin6.sin6_addr, &b->addr.sin6.sin6_addr, 
-	    sizeof(struct in6_addr)) == 0;
- default: /* fall through and signal error */
-   ;
- }
- return 0;
-}
-
-#else /* WITH_POSIX */
-
-/* make compilers happy that do not like empty modules */
-static inline void dummy()
-{
-}
-
-#endif /* not WITH_POSIX */
-
